+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Fixed bugs
 ~~~~~~~~~~~
 
 - Inactive questions could not be deleted (making them active first worked as a workaround). (#289)
+- Choice questions could not be deleted as long as they still had answer options. (#288)
 
 
 .. _releases: https://github.com/pretalx/pretalx/releases